requirements.txt
wheel
asyncpg
sqlalchemy[asyncio]
typing_extensions
discord.py
click
